+++ apps/cli/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 - fix: Pass `flutter_assets` to backend when using `celest deploy`
+- fix: Sentry integration
 
 ## 1.0.10+2